Updated Readme.md
Acknowledging that as part of completing the assignments other pieces of information for reference may be present. This is the assumption and with the intention that the shared resources are used for reference and thought provocation and not plagiarism or the reuse of any code that is not my own.